+++ xclib/tests/11_config_test.py
@@ -1,3 +1,4 @@
+# Test whether the configuration parser works as it should
 import sys
 import unittest
 from xclib.tests.iostub import iostub
@@ -23,8 +24,8 @@ class TestConfiguration(unittest.TestCase, iostub):
                   '--timeout', '5',
                   '--cache-unreachable-ttl', '1w',
                   '--cache-query-ttl', '3600'])
-        print args.timeout
-        assert args.timeout == 5
+        print(args.timeout)
+        self.assertEqual(args.timeout, 5)
 
     def test_xcauth_timeout(self):
         args = get_args('/var/log/xcauth', None, None, 'xcauth',
@@ -36,13 +37,13 @@ class TestConfiguration(unittest.TestCase, iostub):
                   '--timeout', '1,2',
                   '--cache-unreachable-ttl', '1w',
                   '--cache-query-ttl', '3600'])
-        print args.timeout
-        assert args.timeout == (1, 2)
+        print(args.timeout)
+        self.assertEqual(args.timeout, (1, 2))
 
     def test_xcauth_crash_timeout(self):
         self.stub_stdouts()
-        try:
-            args = get_args('/var/log/xcauth', None, None, 'xcauth',
+        self.assertRaises(ValueError, get_args, 
+                '/var/log/xcauth', None, None, 'xcauth',
                 config_file_contents='#',
                 args=['-b', '/tmp/domdb.db',
                       '--secret', '012345678',
@@ -52,14 +53,11 @@ class TestConfiguration(unittest.TestCase, iostub):
                       '--timeout', '1,2,3',
                       '--cache-unreachable-ttl', '1w',
                       '--cache-query-ttl', '3600'])
-            assert False # Should raise ValueError
-        except ValueError:
-            pass
 
     def test_xcauth_exit_a(self):
         self.stub_stdouts()
-        try:
-            args = get_args('/var/log/xcauth', None, None, 'xcauth',
+        self.assertRaises(SystemExit, get_args,
+                '/var/log/xcauth', None, None, 'xcauth',
                 config_file_contents='#',
                 args=['-b', '/tmp/domdb.db',
                       '--secret', '012345678',
@@ -68,22 +66,16 @@ class TestConfiguration(unittest.TestCase, iostub):
                       '--type', 'generic',
                       '--cache-unreachable-ttl', '1w',
                       '--cache-query-ttl', '3600'])
-            assert False # Should exit(1)
-        except SystemExit:
-            pass
 
     def test_xcauth_exit_b(self):
         self.stub_stdouts()
-        try:
-            args = get_args('/var/log/xcauth', None, None, 'xcauth',
+        self.assertRaises(SystemExit, get_args,
+                '/var/log/xcauth', None, None, 'xcauth',
                 config_file_contents='#',
                 args=['-b', '/tmp/domdb.db',
                       '--secret', '012345678',
                       '--url', 'https://unconfigured.example.ch',
                       '--cache-query-ttl', '3600'])
-            assert False # Should exit(1)
-        except SystemExit:
-            pass
 
     def test_xcdbm(self):
         args = get_args('/var/log/xcauth', None, None, 'xcdbm',