+++ xclib/tests/20_prosody_test.py
@@ -1,7 +1,9 @@
+# Checks that prosody_io works as designed
 import sys
 import unittest
 from xclib.prosody_io import prosody_io
 from xclib.tests.iostub import iostub
+from xclib.check import assertEqual
 
 class TestProsody(unittest.TestCase, iostub):
 
@@ -9,15 +11,11 @@ class TestProsody(unittest.TestCase, iostub):
         self.stub_stdin('isuser:login:\n' +
             'auth:log:dom:pass\n')
         tester = iter(prosody_io.read_request())
-        output = tester.next()
-        assert output == ('isuser', 'login', '')
-        output = tester.next()
-        assert output == ('auth', 'log', 'dom', 'pass')
-        try:
-            output = tester.next()
-            assert False # Should raise StopIteration
-        except StopIteration:
-            pass
+        output = next(tester)
+        assertEqual(output, ('isuser', 'login', ''))
+        output = next(tester)
+        assertEqual(output, ('auth', 'log', 'dom', 'pass'))
+        self.assertRaises(StopIteration, next, tester)
 
     def test_output_false(self):
         self.stub_stdout()